Working Triton-specfic Error Handling
